Added a SetDN method to AddRequest.
Added a SetDN method to AddRequest. A usecase is, the result returned from AD is of uppercase realm for DN, and if a comparison needs to be done against an openldap result, it has to be converted to something similar in lowercase. This method makes it easy, as the SetDN sets the value to the pointer. Eg: ```i.SetDN(r.ReplaceAllStringFunc(input, func(m string) string {
			return strings.ToLower(m)
		}))```