Fix control packets being encoded when none exist on bind
During the code reshuffle in #126 at bind time a call to encodeControls
was added. Unfortunately, the safety check around calling it when there
are no controls to encode were omitted from this call (unlike in del.go
and search.go). As a result the packet was always getting control
characters added even if there were none to encode.

I also modified the checks in del.go and search.go to be a little safer;
rather than check for a nil slice, it also will do the right thing when
the slice is not nil but there are no entries.