add DialURL function to simplify handling of ldap:// and ldaps:// urls
in some cases a client application wants simply to connect to an ldap
server based on provided ldap url. To simplify client-side code a
convenient function DialURL is added which based on provided url will
use plain tcp or tls connection. If no port number is provided, the
standard ports defined by IANA is used: 389 for plain and 636 for ssl
connections.