Update golang.org/x/net dependency
The vulnerability in golang.org/x/net prior to version 0.36.0 does not
apply to this module, however for general hygiene we're updating the
dependencies.

This has pushed the issue of which versions of Go we support; the go.mod
file from golang.org/x/net requires Go version 1.23 minimum, and given
that the Go maintainers only support the last two versions, this change
reduces the test matrix to the last two versions.